requirements.txt
asgiref==3.8.0
attrs==23.2.0
black==24.3.0
click==8.1.7
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
drf-spectacular==0.27.1
flake8==7.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
uritemplate==4.1.1
